Refactor block building decoder creation to connector specific class
