+++ lib/trino-hive-formats/src/main/java/io/trino/hive/formats/avro/AvroTypeManager.java
@@ -14,36 +14,14 @@
 package io.trino.hive.formats.avro;
 
 import io.trino.spi.block.Block;
-import io.trino.spi.block.BlockBuilder;
 import io.trino.spi.type.Type;
 import org.apache.avro.Schema;
 
-import java.util.Map;
 import java.util.Optional;
-import java.util.function.BiConsumer;
 import java.util.function.BiFunction;
 
 public interface AvroTypeManager
 {
-    /**
-     * Called when the type manager is reading out data from a data file such as in {@link AvroFileReader}
-     *
-     * @param fileMetadata metadata from the file header
-     */
-    void configure(Map<String, byte[]> fileMetadata);
-
-    Optional<Type> overrideTypeForSchema(Schema schema)
-            throws AvroTypeException;
-
-    /**
-     * Object provided by FasterReader's deserialization with no conversions.
-     * Object class determined by Avro's standard generic data process
-     * BlockBuilder provided by Type returned above for the schema
-     * Possible to override for each primitive type as well.
-     */
-    Optional<BiConsumer<BlockBuilder, Object>> overrideBuildingFunctionForSchema(Schema schema)
-            throws AvroTypeException;
-
     /**
      * Extract and convert the object from the given block at the given position and return the Avro Generic Data forum.
      * Type is either provided explicitly to the writer or derived from the schema using this interface.
