+++ lib/trino-hive-formats/src/main/java/io/trino/hive/formats/avro/model/SkipFieldRecordFieldReadAction.java
@@ -13,7 +13,11 @@
  */
 package io.trino.hive.formats.avro.model;
 
-import io.trino.hive.formats.avro.AvroPageDataReader.RowBlockBuildingDecoder.SkipSchemaBuildingAction.SkipAction;
+import org.apache.avro.Schema;
+import org.apache.avro.io.Decoder;
+
+import java.io.IOException;
+import java.util.List;
 
 import static java.util.Objects.requireNonNull;
 
@@ -23,5 +27,124 @@ public record SkipFieldRecordFieldReadAction(SkipAction skipAction)
     public SkipFieldRecordFieldReadAction
     {
         requireNonNull(skipAction, "skipAction is null");
+    }
+
+    @FunctionalInterface
+    public interface SkipAction
+    {
+        void skip(Decoder decoder)
+                throws IOException;
+    }
+
+    public static SkipAction createSkipActionForSchema(Schema schema)
+    {
+        return switch (schema.getType()) {
+            case NULL -> Decoder::readNull;
+            case BOOLEAN -> Decoder::readBoolean;
+            case INT -> Decoder::readInt;
+            case LONG -> Decoder::readLong;
+            case FLOAT -> Decoder::readFloat;
+            case DOUBLE -> Decoder::readDouble;
+            case STRING -> Decoder::skipString;
+            case BYTES -> Decoder::skipBytes;
+            case ENUM -> Decoder::readEnum;
+            case FIXED -> {
+                int size = schema.getFixedSize();
+                yield decoder -> decoder.skipFixed(size);
+            }
+            case ARRAY -> new ArraySkipAction(schema.getElementType());
+            case MAP -> new MapSkipAction(schema.getValueType());
+            case RECORD -> new RecordSkipAction(schema.getFields());
+            case UNION -> new UnionSkipAction(schema.getTypes());
+        };
+    }
+
+    private static class ArraySkipAction
+            implements SkipAction
+    {
+        private final SkipAction elementSkipAction;
+
+        public ArraySkipAction(Schema elementSchema)
+        {
+            elementSkipAction = createSkipActionForSchema(requireNonNull(elementSchema, "elementSchema is null"));
+        }
+
+        @Override
+        public void skip(Decoder decoder)
+                throws IOException
+        {
+            for (long i = decoder.skipArray(); i != 0; i = decoder.skipArray()) {
+                for (long j = 0; j < i; j++) {
+                    elementSkipAction.skip(decoder);
+                }
+            }
+        }
+    }
+
+    private static class MapSkipAction
+            implements SkipFieldRecordFieldReadAction.SkipAction
+    {
+        private final SkipFieldRecordFieldReadAction.SkipAction valueSkipAction;
+
+        public MapSkipAction(Schema valueSchema)
+        {
+            valueSkipAction = createSkipActionForSchema(requireNonNull(valueSchema, "valueSchema is null"));
+        }
+
+        @Override
+        public void skip(Decoder decoder)
+                throws IOException
+        {
+            for (long i = decoder.skipMap(); i != 0; i = decoder.skipMap()) {
+                for (long j = 0; j < i; j++) {
+                    decoder.skipString(); // key
+                    valueSkipAction.skip(decoder); // value
+                }
+            }
+        }
+    }
+
+    private static class RecordSkipAction
+            implements SkipFieldRecordFieldReadAction.SkipAction
+    {
+        private final SkipFieldRecordFieldReadAction.SkipAction[] fieldSkips;
+
+        public RecordSkipAction(List<Schema.Field> fields)
+        {
+            fieldSkips = new SkipFieldRecordFieldReadAction.SkipAction[requireNonNull(fields, "fields is null").size()];
+            for (int i = 0; i < fields.size(); i++) {
+                fieldSkips[i] = createSkipActionForSchema(fields.get(i).schema());
+            }
+        }
+
+        @Override
+        public void skip(Decoder decoder)
+                throws IOException
+        {
+            for (SkipFieldRecordFieldReadAction.SkipAction fieldSkipAction : fieldSkips) {
+                fieldSkipAction.skip(decoder);
+            }
+        }
+    }
+
+    private static class UnionSkipAction
+            implements SkipFieldRecordFieldReadAction.SkipAction
+    {
+        private final SkipFieldRecordFieldReadAction.SkipAction[] skipActions;
+
+        private UnionSkipAction(List<Schema> types)
+        {
+            skipActions = new SkipFieldRecordFieldReadAction.SkipAction[requireNonNull(types, "types is null").size()];
+            for (int i = 0; i < types.size(); i++) {
+                skipActions[i] = createSkipActionForSchema(types.get(i));
+            }
+        }
+
+        @Override
+        public void skip(Decoder decoder)
+                throws IOException
+        {
+            skipActions[decoder.readIndex()].skip(decoder);
+        }
     }
 }
