+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/avro/AvroHiveFileUtils.java
@@ -45,17 +45,17 @@ import java.util.Optional;
 import java.util.concurrent.atomic.AtomicInteger;
 
 import static com.google.common.collect.ImmutableMap.toImmutableMap;
+import static io.trino.hive.formats.avro.AvroHiveConstants.CHAR_TYPE_LOGICAL_NAME;
+import static io.trino.hive.formats.avro.AvroHiveConstants.SCHEMA_DOC;
+import static io.trino.hive.formats.avro.AvroHiveConstants.SCHEMA_LITERAL;
+import static io.trino.hive.formats.avro.AvroHiveConstants.SCHEMA_NAME;
+import static io.trino.hive.formats.avro.AvroHiveConstants.SCHEMA_NAMESPACE;
+import static io.trino.hive.formats.avro.AvroHiveConstants.SCHEMA_NONE;
+import static io.trino.hive.formats.avro.AvroHiveConstants.SCHEMA_URL;
+import static io.trino.hive.formats.avro.AvroHiveConstants.TABLE_NAME;
+import static io.trino.hive.formats.avro.AvroHiveConstants.VARCHAR_AND_CHAR_LOGICAL_TYPE_LENGTH_PROP;
+import static io.trino.hive.formats.avro.AvroHiveConstants.VARCHAR_TYPE_LOGICAL_NAME;
 import static io.trino.plugin.hive.HiveMetadata.TABLE_COMMENT;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.CHAR_TYPE_LOGICAL_NAME;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.SCHEMA_DOC;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.SCHEMA_LITERAL;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.SCHEMA_NAME;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.SCHEMA_NAMESPACE;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.SCHEMA_NONE;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.SCHEMA_URL;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.TABLE_NAME;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.VARCHAR_AND_CHAR_LOGICAL_TYPE_LENGTH_PROP;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.VARCHAR_TYPE_LOGICAL_NAME;
 import static io.trino.plugin.hive.util.HiveUtil.getColumnNames;
 import static io.trino.plugin.hive.util.HiveUtil.getColumnTypes;
 import static io.trino.plugin.hive.util.SerdeConstants.LIST_COLUMN_COMMENTS;
