+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/avro/TestAvroSchemaGeneration.java
@@ -29,9 +29,9 @@ import java.util.Properties;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import static io.trino.hive.formats.avro.AvroHiveConstants.TABLE_NAME;
 import static io.trino.plugin.hive.HiveType.HIVE_STRING;
 import static io.trino.plugin.hive.HiveType.toHiveType;
-import static io.trino.plugin.hive.avro.AvroHiveConstants.TABLE_NAME;
 import static io.trino.plugin.hive.util.SerdeConstants.LIST_COLUMNS;
 import static io.trino.plugin.hive.util.SerdeConstants.LIST_COLUMN_TYPES;
 import static io.trino.spi.type.VarcharType.VARCHAR;
